# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WeWorkout API",
        "description": "Documentación de la API del backend de la aplicación WeWorkout. La mayoría de llamadas utilizan autenticación mediante JWT. Se incluyen ejemplos de request y de response.",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/workouts": {
            "get": {
                "tags": [
                    "Registros de ejercicio"
                ],
                "summary": "Obtener registros de ejercicios",
                "operationId": "App\\Http\\Controllers\\WorkoutController::index",
                "responses": {
                    "200": {
                        "description": "Obtener todos los registros de ejercicio.",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 200,
                                    "message": 200,
                                    "data": [
                                        {
                                            "id": 17,
                                            "created_at": "2021-05-31T21:11:09.000000Z",
                                            "updated_at": "2021-05-31T21:11:09.000000Z",
                                            "name": "Earum aliquam quos illo error qui.",
                                            "date": "2020-05-17 15:27:54",
                                            "weight": 50,
                                            "fat_percentage": 80,
                                            "notes": "Consequatur animi architecto ut ut. Saepe mollitia quo doloremque amet ex et aspernatur voluptatibus. In et doloremque et magni hic. Voluptatem quaerat reiciendis nostrum optio veniam hic aut.",
                                            "user_id": 1,
                                            "sets": [
                                                {
                                                    "id": 256,
                                                    "created_at": "2021-05-31T21:11:08.000000Z",
                                                    "updated_at": "2021-05-31T21:11:08.000000Z",
                                                    "repetitions": 16,
                                                    "exercise_id": 9,
                                                    "pivot": {
                                                        "workout_id": 17,
                                                        "set_id": 256,
                                                        "weight": 71
                                                    }
                                                }
                                            ]
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Registros de ejercicio"
                ],
                "summary": "Crear Registro de ejercicio",
                "operationId": "App\\Http\\Controllers\\WorkoutController::store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "user_id": 1,
                                    "name": "Workout SETS",
                                    "notes": "Notas de ejemplo",
                                    "weight": 50,
                                    "fat_percentage": 42,
                                    "date": "2021-03-21 14:14:36",
                                    "sets": [
                                        {
                                            "exercise": 1,
                                            "repetitions": 10,
                                            "weight": 22
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK. Crea una registro de ejercicio",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 201,
                                    "message": "Workout creado con éxito",
                                    "data": {
                                        "user_id": 1,
                                        "name": "Workout Test",
                                        "notes": "Workout Notes",
                                        "weight": 50,
                                        "fat_percentage": 42,
                                        "date": "2021-03-21 14:14:36",
                                        "updated_at": "2021-06-01T08:57:51.000000Z",
                                        "created_at": "2021-06-01T08:57:51.000000Z",
                                        "id": 751,
                                        "sets": [
                                            {
                                                "id": 10,
                                                "created_at": "2021-05-31T21:11:08.000000Z",
                                                "updated_at": "2021-05-31T21:11:08.000000Z",
                                                "repetitions": 10,
                                                "exercise_id": 1,
                                                "pivot": {
                                                    "workout_id": 751,
                                                    "set_id": 10,
                                                    "weight": 22
                                                }
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "400": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error añadiendo registro de ejercicio"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/workouts/{id}": {
            "put": {
                "tags": [
                    "Registros de ejercicio"
                ],
                "summary": "Actualizar Registro de ejercicio",
                "operationId": "App\\Http\\Controllers\\WorkoutController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del registro de ejercicio",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "user_id": 1,
                                    "name": "Workout SETS",
                                    "notes": "Notas de ejemplo",
                                    "weight": 50,
                                    "fat_percentage": 42,
                                    "date": "2021-03-21 14:14:36",
                                    "sets": [
                                        {
                                            "exercise": 1,
                                            "repetitions": 10,
                                            "weight": 22
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK. Actualiza un registro de ejercicio",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 201,
                                    "message": "Workout editado con éxito",
                                    "data": {
                                        "id": 22,
                                        "created_at": "2021-05-31T21:11:09.000000Z",
                                        "updated_at": "2021-06-01T09:33:47.000000Z",
                                        "name": "Workout Test",
                                        "date": "2021-03-21 14:14:36",
                                        "weight": 50,
                                        "fat_percentage": 42,
                                        "notes": "Update test",
                                        "user_id": 1,
                                        "sets": [
                                            {
                                                "id": 10,
                                                "created_at": "2021-05-31T21:11:08.000000Z",
                                                "updated_at": "2021-05-31T21:11:08.000000Z",
                                                "repetitions": 10,
                                                "exercise_id": 1,
                                                "pivot": {
                                                    "workout_id": 22,
                                                    "set_id": 10,
                                                    "weight": 22
                                                }
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "400": {
                        "description": "Error de validación "
                    },
                    "500": {
                        "description": "Error actualizando registro de ejercicio"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Registros de ejercicio"
                ],
                "summary": "Borrar registro de ejerccio",
                "operationId": "App\\Http\\Controllers\\WorkoutController::delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del registro de ejercicio",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK. Borra un registro de ejercicio",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 200,
                                    "message": "Workout eliminado con éxito",
                                    "data": {
                                        "id": 17
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}